Very fast AVX2 implementation
